[DEV-2] separate Room Database module
