+++ app/src/main/java/com/fakhry/lifelog/ui/activities/read/ReadViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.fakhry.lifelog.data.Repository
-import com.fakhry.lifelog.data.local.entities.NoteEntity
-import com.fakhry.lifelog.data.local.relation.NoteWithEditLogsRelation
-import com.fakhry.lifelog.data.local.relation.NoteWithTagRelation
+import com.fakhry.lifelog.core.data.repository.Repository
+import com.fakhry.lifelog.storage.model.NoteEntity
+import com.fakhry.lifelog.storage.model.relation.NoteWithEditLogsRelation
+import com.fakhry.lifelog.storage.model.relation.NoteWithTagRelation
 import kotlinx.coroutines.launch
 
 class ReadViewModel(private val mRepository: Repository) : ViewModel() {
