+++ app/src/main/java/com/fakhry/lifelog/ui/adapters/StaggeredTagAdapter.kt
@@ -3,7 +3,7 @@ package com.fakhry.lifelog.ui.adapters
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.fakhry.lifelog.data.local.entities.TagEntity
+import com.fakhry.lifelog.storage.model.TagEntity
 import com.fakhry.lifelog.databinding.ItemRowTagBinding
 
 class StaggeredTagAdapter : RecyclerView.Adapter<StaggeredTagAdapter.GridViewHolder>() {
