+++ app/src/main/java/com/fakhry/lifelog/ui/fragments/calendar/CalendarFragment.kt
@@ -6,10 +6,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.observe
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.fakhry.lifelog.base.BaseFunction
-import com.fakhry.lifelog.data.local.entities.NoteEntity
+import com.fakhry.lifelog.storage.model.NoteEntity
 import com.fakhry.lifelog.databinding.FragmentCalendarBinding
 import com.fakhry.lifelog.ui.adapters.ListNoteAdapter
 import com.fakhry.lifelog.viewmodel.ViewModelFactory
