+++ app/src/main/java/com/fakhry/lifelog/ui/fragments/favorite/FavoriteFragment.kt
@@ -6,10 +6,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.observe
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.fakhry.lifelog.data.local.entities.DateNoteEntity
-import com.fakhry.lifelog.data.local.entities.NoteEntity
+import com.fakhry.lifelog.storage.model.DateNoteEntity
+import com.fakhry.lifelog.storage.model.NoteEntity
 import com.fakhry.lifelog.databinding.FragmentFavoriteBinding
 import com.fakhry.lifelog.ui.adapters.ListDateWithNoteAdapter
 import com.fakhry.lifelog.viewmodel.ViewModelFactory
