+++ settings.gradle.kts
@@ -22,8 +22,11 @@ dependencyResolutionManagement {
 
 rootProject.name = "LifeLog - Daily Note and Mood Tracker"
 include(":app")
+include(":core")
 include(":storage")
 include(":features:settings")
 include(":commons:utils")
 include(":commons:resources")
 include(":commons:navigation")
+include(":features:settings")
+include(":features:favorites")
