+++ storage/src/main/java/com/fakhry/lifelog/storage/model/relation/NoteWithEditLogsRelation.kt
@@ -1,9 +1,9 @@
-package com.fakhry.lifelog.data.local.relation
+package com.fakhry.lifelog.storage.model.relation
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import com.fakhry.lifelog.data.local.entities.EditLogEntity
-import com.fakhry.lifelog.data.local.entities.NoteEntity
+import com.fakhry.lifelog.storage.model.EditLogEntity
+import com.fakhry.lifelog.storage.model.NoteEntity
 
 data class NoteWithEditLogsRelation(
     @Embedded
