+++ storage/src/main/java/com/fakhry/lifelog/storage/model/relation/TagWithNoteRelation.kt
@@ -1,10 +1,10 @@
-package com.fakhry.lifelog.data.local.relation
+package com.fakhry.lifelog.storage.model.relation
 
 import androidx.room.Embedded
 import androidx.room.Junction
 import androidx.room.Relation
-import com.fakhry.lifelog.data.local.entities.NoteEntity
-import com.fakhry.lifelog.data.local.entities.TagEntity
+import com.fakhry.lifelog.storage.model.NoteEntity
+import com.fakhry.lifelog.storage.model.TagEntity
 
 data class TagWithNoteRelation(
     @Embedded
