+++ storage/src/main/java/com/fakhry/lifelog/storage/room/LifeLogDao.kt
@@ -1,13 +1,13 @@
-package com.fakhry.lifelog.data.local.room
+package com.fakhry.lifelog.storage.room
 
 import androidx.room.*
-import com.fakhry.lifelog.data.local.entities.EditLogEntity
-import com.fakhry.lifelog.data.local.entities.NoteEntity
-import com.fakhry.lifelog.data.local.entities.TagEntity
-import com.fakhry.lifelog.data.local.relation.NoteTagCrossRef
-import com.fakhry.lifelog.data.local.relation.NoteWithEditLogsRelation
-import com.fakhry.lifelog.data.local.relation.NoteWithTagRelation
-import com.fakhry.lifelog.data.local.relation.TagWithNoteRelation
+import com.fakhry.lifelog.storage.model.EditLogEntity
+import com.fakhry.lifelog.storage.model.NoteEntity
+import com.fakhry.lifelog.storage.model.TagEntity
+import com.fakhry.lifelog.storage.model.relation.NoteTagCrossRef
+import com.fakhry.lifelog.storage.model.relation.NoteWithEditLogsRelation
+import com.fakhry.lifelog.storage.model.relation.NoteWithTagRelation
+import com.fakhry.lifelog.storage.model.relation.TagWithNoteRelation
 
 @Dao
 interface LifeLogDao {
